pyproject.toml
[project]
name = "swarm-saver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.40.16",
    "geopandas>=1.1.1",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
